modularized ports, volumes and depends on options into nodes and links component
