+++ src/renderer/components/Links.tsx
@@ -11,15 +11,16 @@
 import React, { useEffect } from 'react';
 import * as d3 from 'd3';
 // IMPORT HELPER FUNCTIONS
-import { Link, Services } from '../App.d';
+import { Link, Services, Options } from '../App.d';
 // IMPORT TYPES
 
 type Props = {
   links: Link[];
   services: Services;
+  options: Options;
 };
 
-const Links: React.FC<Props> = ({ links, services }) => {
+const Links: React.FC<Props> = ({ links, services, options }) => {
   useEffect(() => {
     //initialize graph
     const forceGraph = d3.select('.graph');
@@ -62,6 +63,21 @@ const Links: React.FC<Props> = ({ links, services }) => {
       arrowsGroup.remove();
     };
   }, [services]);
+
+  /**
+   *********************
+   * DEPENDS ON OPTION TOGGLE
+   *********************
+   */
+  useEffect(() => {
+    if (options.dependsOn) {
+      d3.select('.arrowsGroup').classed('hide', false);
+      d3.select('.links').classed('hide', false);
+    } else {
+      d3.select('.arrowsGroup').classed('hide', true);
+      d3.select('.links').classed('hide', true);
+    }
+  }, [options.dependsOn]);
 
   return <g className="links"></g>;
 };
