--- src/renderer/components/Ports.tsx
@@ -1,19 +1,0 @@
-/**
- * ************************************
- *
- * @module  Ports.tsx
- * @author
- * @date 3/23/20
- * @description Rendering of the ports in d3 simulation
- *
- * ************************************
- */
-import React from 'react';
-
-type Props = {};
-
-const Ports: React.FC<Props> = props => {
-  return <></>;
-};
-
-export default Ports;
