--- src/renderer/components/Service.tsx
@@ -1,28 +1,0 @@
-/**
- * ************************************
- *
- * @module  Service.tsx
- * @author
- * @date 3/11/20
- * @description display a single service container
- *
- * ************************************
- */
-import React from 'react';
-
-import { Service } from '../App.d';
-
-type Props = {
-  name: string;
-  service: Service;
-};
-
-const Service: React.FC<Props> = ({ service }) => {
-  return (
-    <div>
-      <div></div>
-    </div>
-  );
-};
-
-export default Service;
